refactor(session): extract the snapshot diff engine into internal/session/snapdiff
diffErrors, diffNetwork, diffPerformance and computeSummary were declared as
methods on *SessionManager, which is why the filenames and the package layout
both said they belonged next to the manager. None of the four reads the
receiver. They are pure functions of two *NamedSnapshot values, and the tests
were the tell: every one of them built a mock CaptureStateReader and a
SessionManager purely to have something to call the method on, then never
touched it again.

Turned them into package-level functions in a leaf package that knows nothing
about snapshot storage:

  snapdiff.Errors(a, b)      (was sm.diffErrors)
  snapdiff.Network(a, b)     (was sm.diffNetwork)
  snapdiff.Performance(a, b) (was sm.diffPerformance)
  snapdiff.Summarize(result) (was sm.computeSummary)

SessionManager.Compare still resolves the two operands (including the "current"
live-state case) and now delegates the arithmetic.

The types shed their Session prefix now that the package name carries it —
SessionDiffResult is snapdiff.Result, SessionNetworkDiff is snapdiff.NetworkDiff,
DiffSummary is snapdiff.Summary. Every JSON tag is unchanged, so the
diff_sessions response is byte-identical. type_aliases.go loses the DiffResult /
NetworkDiff / NetworkChange aliases it carried to hide that stutter; they had no
users. Manager = SessionManager stays, it has three.

actions_diff_test.go was two test files in a trench coat: diff assertions plus
validateName/removeFromOrder/Compare-integration tests that belong to the
manager. Split accordingly; internal/session/helpers_test.go holds the latter.

  internal/session           18 -> 14 files
  internal/session/snapdiff   4 files (new), every function 100% covered
