+++ docs/features/feature/historical-snapshots/index.md
@@ -4,13 +4,22 @@ feature_id: feature-historical-snapshots
 status: proposed
 feature_type: feature
 owners: []
-last_reviewed: 2026-07-05
+last_reviewed: 2026-07-26
 code_paths:
   - internal/types/snapshot.go
   - internal/session/types.go
   - internal/session/snapshot-manager.go
+  - internal/session/comparison.go
+  - internal/session/snapdiff/types.go
+  - internal/session/snapdiff/errors.go
+  - internal/session/snapdiff/network.go
+  - internal/session/snapdiff/performance.go
 test_paths:
   - internal/session/snapshot_manager_test.go
+  - internal/session/comparison_test.go
+  - internal/session/snapdiff/errors_test.go
+  - internal/session/snapdiff/network_test.go
+  - internal/session/snapdiff/performance_test.go
 last_verified_version: 0.7.12
 last_verified_date: 2026-03-05
 ---
@@ -42,5 +51,8 @@ last_verified_date: 2026-03-05
   - `internal/types/snapshot.go`
   - `internal/session/types.go`
   - `internal/session/snapshot-manager.go`
+- Snapshot diff engine:
+  - `internal/session/snapdiff/`
 - Tests:
   - `internal/session/snapshot_manager_test.go`
+  - `internal/session/snapdiff/`
