+++ internal/session/snapdiff/errors_test.go
@@ -1,17 +1,11 @@
 // Purpose: Tests for session actions diff computation.
 // Docs: docs/features/feature/pagination/index.md
 
-// actions_diff_test.go — Tests for actions-diff.go and helpers.go.
-// Covers: diffErrors, countPerfRegressions, hasStatusRegression, computeSummary,
-// validateName, removeFromOrder.
-package session
+// errors_test.go — Tests for errors.go.
+// Covers: Errors, countPerfRegressions, hasStatusRegression, Summarize.
+package snapdiff
 
-import (
-	"strings"
-	"testing"
-
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/performance"
-)
+import "testing"
 
 // ============================================
 // diffErrors
@@ -19,16 +13,13 @@ import (
 
 func TestDiffErrors_NewErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	snapA := &NamedSnapshot{ConsoleErrors: []SnapshotError{}}
 	snapB := &NamedSnapshot{ConsoleErrors: []SnapshotError{
 		{Type: "error", Message: "TypeError: x is null", Count: 2},
 		{Type: "error", Message: "RangeError: invalid index", Count: 1},
 	}}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.New) != 2 {
 		t.Fatalf("Expected 2 new errors, got %d", len(diff.New))
@@ -43,16 +34,13 @@ func TestDiffErrors_NewErrors(t *testing.T) {
 
 func TestDiffErrors_ResolvedErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	snapA := &NamedSnapshot{ConsoleErrors: []SnapshotError{
 		{Type: "error", Message: "old error 1", Count: 1},
 		{Type: "error", Message: "old error 2", Count: 3},
 	}}
 	snapB := &NamedSnapshot{ConsoleErrors: []SnapshotError{}}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.Resolved) != 2 {
 		t.Fatalf("Expected 2 resolved errors, got %d", len(diff.Resolved))
@@ -67,16 +55,13 @@ func TestDiffErrors_ResolvedErrors(t *testing.T) {
 
 func TestDiffErrors_UnchangedErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	errors := []SnapshotError{
 		{Type: "error", Message: "persistent error", Count: 5},
 	}
 	snapA := &NamedSnapshot{ConsoleErrors: errors}
 	snapB := &NamedSnapshot{ConsoleErrors: errors}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.Unchanged) != 1 {
 		t.Fatalf("Expected 1 unchanged error, got %d", len(diff.Unchanged))
@@ -97,9 +82,6 @@ func TestDiffErrors_UnchangedErrors(t *testing.T) {
 
 func TestDiffErrors_MixedChanges(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	snapA := &NamedSnapshot{ConsoleErrors: []SnapshotError{
 		{Type: "error", Message: "stays", Count: 1},
 		{Type: "error", Message: "goes away", Count: 1},
@@ -109,7 +91,7 @@ func TestDiffErrors_MixedChanges(t *testing.T) {
 		{Type: "error", Message: "brand new", Count: 1},
 	}}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.New) != 1 {
 		t.Errorf("Expected 1 new error, got %d", len(diff.New))
@@ -135,13 +117,10 @@ func TestDiffErrors_MixedChanges(t *testing.T) {
 
 func TestDiffErrors_BothEmpty(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	snapA := &NamedSnapshot{ConsoleErrors: []SnapshotError{}}
 	snapB := &NamedSnapshot{ConsoleErrors: []SnapshotError{}}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.New) != 0 {
 		t.Errorf("Expected 0 new, got %d", len(diff.New))
@@ -156,13 +135,10 @@ func TestDiffErrors_BothEmpty(t *testing.T) {
 
 func TestDiffErrors_NilConsoleErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	snapA := &NamedSnapshot{ConsoleErrors: nil}
 	snapB := &NamedSnapshot{ConsoleErrors: nil}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	if len(diff.New) != 0 {
 		t.Errorf("Expected 0 new for nil, got %d", len(diff.New))
@@ -174,9 +150,6 @@ func TestDiffErrors_NilConsoleErrors(t *testing.T) {
 
 func TestDiffErrors_DuplicateMessagesDeduped(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
 	// Two entries with same message in A - map deduplication means last wins
 	snapA := &NamedSnapshot{ConsoleErrors: []SnapshotError{
 		{Type: "error", Message: "duplicate", Count: 1},
@@ -184,7 +157,7 @@ func TestDiffErrors_DuplicateMessagesDeduped(t *testing.T) {
 	}}
 	snapB := &NamedSnapshot{ConsoleErrors: []SnapshotError{}}
 
-	diff := sm.diffErrors(snapA, snapB)
+	diff := Errors(snapA, snapB)
 
 	// Should be 1 resolved (deduped by message)
 	if len(diff.Resolved) != 1 {
@@ -250,7 +223,7 @@ func TestCountPerfRegressions_PartialNilMixed(t *testing.T) {
 
 func TestHasStatusRegression_OKToError(t *testing.T) {
 	t.Parallel()
-	changes := []SessionNetworkChange{
+	changes := []NetworkChange{
 		{BeforeStatus: 200, AfterStatus: 500},
 	}
 	if !hasStatusRegression(changes) {
@@ -260,7 +233,7 @@ func TestHasStatusRegression_OKToError(t *testing.T) {
 
 func TestHasStatusRegression_ErrorToOK(t *testing.T) {
 	t.Parallel()
-	changes := []SessionNetworkChange{
+	changes := []NetworkChange{
 		{BeforeStatus: 500, AfterStatus: 200},
 	}
 	if hasStatusRegression(changes) {
@@ -270,7 +243,7 @@ func TestHasStatusRegression_ErrorToOK(t *testing.T) {
 
 func TestHasStatusRegression_ErrorToError(t *testing.T) {
 	t.Parallel()
-	changes := []SessionNetworkChange{
+	changes := []NetworkChange{
 		{BeforeStatus: 400, AfterStatus: 500},
 	}
 	if hasStatusRegression(changes) {
@@ -283,14 +256,14 @@ func TestHasStatusRegression_EmptyChanges(t *testing.T) {
 	if hasStatusRegression(nil) {
 		t.Error("Expected false for nil changes")
 	}
-	if hasStatusRegression([]SessionNetworkChange{}) {
+	if hasStatusRegression([]NetworkChange{}) {
 		t.Error("Expected false for empty changes")
 	}
 }
 
 func TestHasStatusRegression_MultipleChanges(t *testing.T) {
 	t.Parallel()
-	changes := []SessionNetworkChange{
+	changes := []NetworkChange{
 		{BeforeStatus: 500, AfterStatus: 200}, // improvement
 		{BeforeStatus: 200, AfterStatus: 404}, // regression
 	}
@@ -302,11 +275,11 @@ func TestHasStatusRegression_MultipleChanges(t *testing.T) {
 func TestHasStatusRegression_BoundaryValues(t *testing.T) {
 	t.Parallel()
 	// 399 -> 400: OK to error
-	if !hasStatusRegression([]SessionNetworkChange{{BeforeStatus: 399, AfterStatus: 400}}) {
+	if !hasStatusRegression([]NetworkChange{{BeforeStatus: 399, AfterStatus: 400}}) {
 		t.Error("Expected true for 399->400")
 	}
 	// 400 -> 399: error to OK, not a regression in the "OK to error" sense
-	if hasStatusRegression([]SessionNetworkChange{{BeforeStatus: 400, AfterStatus: 399}}) {
+	if hasStatusRegression([]NetworkChange{{BeforeStatus: 400, AfterStatus: 399}}) {
 		t.Error("Expected false for 400->399 (was already erroring, now OK)")
 	}
 }
@@ -317,16 +290,13 @@ func TestHasStatusRegression_BoundaryValues(t *testing.T) {
 
 func TestComputeSummary_Unchanged(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors:      ErrorDiff{New: []SnapshotError{}, Resolved: []SnapshotError{}, Unchanged: []SnapshotError{{Message: "still here"}}},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 
 	if summary.Verdict != "unchanged" {
 		t.Errorf("Expected 'unchanged', got %q", summary.Verdict)
@@ -341,16 +311,13 @@ func TestComputeSummary_Unchanged(t *testing.T) {
 
 func TestComputeSummary_Improved(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors:      ErrorDiff{New: []SnapshotError{}, Resolved: []SnapshotError{{Message: "fixed"}}, Unchanged: []SnapshotError{}},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "improved" {
 		t.Errorf("Expected 'improved', got %q", summary.Verdict)
 	}
@@ -361,16 +328,13 @@ func TestComputeSummary_Improved(t *testing.T) {
 
 func TestComputeSummary_Regressed_NewErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors:      ErrorDiff{New: []SnapshotError{{Message: "new err"}}, Resolved: []SnapshotError{}, Unchanged: []SnapshotError{}},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "regressed" {
 		t.Errorf("Expected 'regressed', got %q", summary.Verdict)
 	}
@@ -378,16 +342,13 @@ func TestComputeSummary_Regressed_NewErrors(t *testing.T) {
 
 func TestComputeSummary_Regressed_NetworkErrors(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors:      ErrorDiff{New: []SnapshotError{}, Resolved: []SnapshotError{}, Unchanged: []SnapshotError{}},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{{Status: 500}}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{{Status: 500}}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "regressed" {
 		t.Errorf("Expected 'regressed', got %q", summary.Verdict)
 	}
@@ -398,16 +359,13 @@ func TestComputeSummary_Regressed_NetworkErrors(t *testing.T) {
 
 func TestComputeSummary_Regressed_PerfRegressions(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors:      ErrorDiff{New: []SnapshotError{}, Resolved: []SnapshotError{}, Unchanged: []SnapshotError{}},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{LoadTime: &MetricChange{Regression: true}},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "regressed" {
 		t.Errorf("Expected 'regressed', got %q", summary.Verdict)
 	}
@@ -418,19 +376,16 @@ func TestComputeSummary_Regressed_PerfRegressions(t *testing.T) {
 
 func TestComputeSummary_Regressed_StatusRegression(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors: ErrorDiff{New: []SnapshotError{}, Resolved: []SnapshotError{}, Unchanged: []SnapshotError{}},
-		Network: SessionNetworkDiff{
+		Network: NetworkDiff{
 			NewErrors:     []SnapshotNetworkRequest{},
-			StatusChanges: []SessionNetworkChange{{BeforeStatus: 200, AfterStatus: 500}},
+			StatusChanges: []NetworkChange{{BeforeStatus: 200, AfterStatus: 500}},
 		},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "regressed" {
 		t.Errorf("Expected 'regressed', got %q", summary.Verdict)
 	}
@@ -438,183 +393,17 @@ func TestComputeSummary_Regressed_StatusRegression(t *testing.T) {
 
 func TestComputeSummary_Mixed(t *testing.T) {
 	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	result := &SessionDiffResult{
+	result := &Result{
 		Errors: ErrorDiff{
 			New:      []SnapshotError{{Message: "new"}},
 			Resolved: []SnapshotError{{Message: "fixed"}},
 		},
-		Network:     SessionNetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []SessionNetworkChange{}},
+		Network:     NetworkDiff{NewErrors: []SnapshotNetworkRequest{}, StatusChanges: []NetworkChange{}},
 		Performance: PerformanceDiff{},
 	}
 
-	summary := sm.computeSummary(result)
+	summary := Summarize(result)
 	if summary.Verdict != "mixed" {
 		t.Errorf("Expected 'mixed', got %q", summary.Verdict)
-	}
-}
-
-// ============================================
-// validateName
-// ============================================
-
-func TestValidateName_EmptyName(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	err := sm.validateName("")
-	if err == nil {
-		t.Fatal("Expected error for empty name")
-	}
-	if !strings.Contains(err.Error(), "empty") {
-		t.Errorf("Error should mention empty: %v", err)
-	}
-}
-
-func TestValidateName_ReservedName(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	err := sm.validateName("current")
-	if err == nil {
-		t.Fatal("Expected error for reserved name 'current'")
-	}
-	if !strings.Contains(err.Error(), "reserved") {
-		t.Errorf("Error should mention reserved: %v", err)
-	}
-}
-
-func TestValidateName_TooLong(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	longName := strings.Repeat("a", maxSnapshotNameLen+1)
-	err := sm.validateName(longName)
-	if err == nil {
-		t.Fatal("Expected error for name exceeding max length")
-	}
-	if !strings.Contains(err.Error(), "50") {
-		t.Errorf("Error should mention max length: %v", err)
-	}
-}
-
-func TestValidateName_ExactMaxLength(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	exactName := strings.Repeat("x", maxSnapshotNameLen)
-	err := sm.validateName(exactName)
-	if err != nil {
-		t.Errorf("Expected no error for name at exact max length, got: %v", err)
-	}
-}
-
-func TestValidateName_ValidName(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	err := sm.validateName("my-snapshot-v2")
-	if err != nil {
-		t.Errorf("Expected no error for valid name, got: %v", err)
-	}
-}
-
-// ============================================
-// removeFromOrder (indirect test through Delete)
-// ============================================
-
-func TestRemoveFromOrder_FirstElement(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	sm.Capture("a", "")
-	sm.Capture("b", "")
-	sm.Capture("c", "")
-
-	sm.Delete("a")
-	list := sm.List()
-
-	if len(list) != 2 {
-		t.Fatalf("Expected 2, got %d", len(list))
-	}
-	if list[0].Name != "b" || list[1].Name != "c" {
-		t.Errorf("Expected [b, c], got [%s, %s]", list[0].Name, list[1].Name)
-	}
-}
-
-func TestRemoveFromOrder_LastElement(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	sm.Capture("a", "")
-	sm.Capture("b", "")
-	sm.Capture("c", "")
-
-	sm.Delete("c")
-	list := sm.List()
-
-	if len(list) != 2 {
-		t.Fatalf("Expected 2, got %d", len(list))
-	}
-	if list[0].Name != "a" || list[1].Name != "b" {
-		t.Errorf("Expected [a, b], got [%s, %s]", list[0].Name, list[1].Name)
-	}
-}
-
-// ============================================
-// Integration: Performance with snapshot manager
-// ============================================
-
-func TestDiffPerformance_IntegrationWithCompare(t *testing.T) {
-	t.Parallel()
-	mock := &mockCaptureState{pageURL: "http://localhost:3000"}
-	sm := NewSessionManager(10, mock)
-
-	mock.performance = &performance.Snapshot{
-		Timing:  performance.Timing{Load: 500},
-		Network: performance.NetworkSummary{RequestCount: 5, TransferSize: 25000},
-	}
-	sm.Capture("baseline", "")
-
-	// Make it much worse
-	mock.performance = &performance.Snapshot{
-		Timing:  performance.Timing{Load: 5000},
-		Network: performance.NetworkSummary{RequestCount: 50, TransferSize: 250000},
-	}
-	sm.Capture("degraded", "")
-
-	diff, err := sm.Compare("baseline", "degraded")
-	if err != nil {
-		t.Fatalf("Compare failed: %v", err)
-	}
-
-	if diff.Performance.LoadTime == nil {
-		t.Fatal("Expected non-nil LoadTime")
-	}
-	if diff.Performance.LoadTime.Before != 500 {
-		t.Errorf("Expected before=500, got %v", diff.Performance.LoadTime.Before)
-	}
-	if diff.Performance.LoadTime.After != 5000 {
-		t.Errorf("Expected after=5000, got %v", diff.Performance.LoadTime.After)
-	}
-	if diff.Performance.LoadTime.Change != "+900%" {
-		t.Errorf("Expected '+900%%', got %q", diff.Performance.LoadTime.Change)
-	}
-	if !diff.Performance.LoadTime.Regression {
-		t.Error("Expected regression=true")
-	}
-
-	// All 3 metrics should have regressed (10x increase)
-	if diff.Summary.PerformanceRegressions != 3 {
-		t.Errorf("Expected 3 performance regressions, got %d", diff.Summary.PerformanceRegressions)
 	}
 }
