+++ internal/session/type_aliases.go
@@ -3,9 +3,10 @@
 
 package session
 
-type (
-	Manager       = SessionManager
-	DiffResult    = SessionDiffResult
-	NetworkDiff   = SessionNetworkDiff
-	NetworkChange = SessionNetworkChange
-)
+// Manager is the non-stuttering spelling of SessionManager used at call sites.
+//
+// The DiffResult / NetworkDiff / NetworkChange aliases that used to live here
+// are gone: the diff types moved to the snapdiff package, where they no longer
+// stutter (snapdiff.Result, snapdiff.NetworkDiff, snapdiff.NetworkChange) and
+// the aliases had no remaining users.
+type Manager = SessionManager
